# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgriPride Mkopo - Agricultural Loan Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Farmer {
  id: number;
  phone: string;
  name: string;
  nationalId: string;
  county: string;
  farmSizeAcres: number;
  cropType: string;
  createdAt: string;
}

export interface FarmerInput {
  phone: string;
  name: string;
  nationalId: string;
  county: string;
  farmSizeAcres: number;
  cropType: string;
}

export interface LoanApplication {
  id: number;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  /** @nullable */
  farmerPhone?: string | null;
  amountRequested: number;
  purpose: string;
  status: string;
  /** @nullable */
  aiScore?: number | null;
  /** @nullable */
  aiRecommendation?: string | null;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface LoanApplicationInput {
  farmerId: number;
  amountRequested: number;
  purpose: string;
}

export interface LoanDecision {
  reviewNote?: string;
  approvedAmount?: number;
}

export interface CreditScoreResult {
  score: number;
  recommendation: string;
  factors: string[];
}

export interface Loan {
  id: number;
  applicationId: number;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  /** @nullable */
  farmerPhone?: string | null;
  amount: number;
  interestRate: number;
  status: string;
  dueDate: string;
  /** @nullable */
  disbursedAt?: string | null;
  /** @nullable */
  mpesaReceipt?: string | null;
  /** @nullable */
  amountPaid?: number | null;
  createdAt: string;
}

export interface DisburseResult {
  message: string;
  conversationId: string;
}

export interface Repayment {
  id: number;
  loanId: number;
  amount: number;
  /** @nullable */
  mpesaReceipt?: string | null;
  paidAt: string;
}

export interface RepaymentInput {
  amount: number;
  phone: string;
}

export interface StkPushResult {
  message: string;
  checkoutRequestId: string;
}

export interface DashboardStats {
  totalFarmers: number;
  totalLoansKes: number;
  activeLoanCount: number;
  defaultRate: number;
  pendingApplications: number;
  disbursedThisMonth: number;
  repaidThisMonth: number;
  portfolioAtRisk: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  farmerName?: string | null;
}

export interface PortfolioSegment {
  label: string;
  count: number;
  amountKes: number;
}

export interface LoanPortfolio {
  byCrop: PortfolioSegment[];
  byCounty: PortfolioSegment[];
  byStatus: PortfolioSegment[];
}

export interface UssdRequest {
  sessionId: string;
  phoneNumber: string;
  text: string;
  networkCode?: string;
  serviceCode?: string;
}

export type MpesaCallbackBody = { [key: string]: unknown };

export interface MpesaCallback {
  Body?: MpesaCallbackBody;
}

export type ListFarmersParams = {
search?: string;
};

export type ListLoanApplicationsParams = {
status?: string;
farmerId?: number;
};

export type ListLoansParams = {
status?: string;
farmerId?: number;
};