# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriPride Mkopo - Agricultural Loan Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farmers:
    get:
      operationId: listFarmers
      tags: [farmers]
      summary: List all farmers
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of farmers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farmer"
    post:
      operationId: createFarmer
      tags: [farmers]
      summary: Register a new farmer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerInput"
      responses:
        "201":
          description: Farmer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
  /farmers/{id}:
    get:
      operationId: getFarmer
      tags: [farmers]
      summary: Get farmer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Farmer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "404":
          description: Not found

  # ─── Loan Applications ───────────────────────────────────────────────────────
  /loan-applications:
    get:
      operationId: listLoanApplications
      tags: [loan-applications]
      summary: List loan applications
      parameters:
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: farmerId
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of loan applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanApplication"
    post:
      operationId: createLoanApplication
      tags: [loan-applications]
      summary: Submit a new loan application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
  /loan-applications/{id}:
    get:
      operationId: getLoanApplication
      tags: [loan-applications]
      summary: Get loan application by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
        "404":
          description: Not found
  /loan-applications/{id}/score:
    post:
      operationId: scoreLoanApplication
      tags: [loan-applications]
      summary: Run AI credit scoring on an application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Credit score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditScoreResult"
  /loan-applications/{id}/approve:
    patch:
      operationId: approveLoanApplication
      tags: [loan-applications]
      summary: Approve a loan application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanDecision"
      responses:
        "200":
          description: Application approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
  /loan-applications/{id}/reject:
    patch:
      operationId: rejectLoanApplication
      tags: [loan-applications]
      summary: Reject a loan application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanDecision"
      responses:
        "200":
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"

  # ─── Loans ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Farmer:
      type: object
      required: [id, phone, name, nationalId, county, farmSizeAcres, cropType, createdAt]
      properties:
        id:            { type: integer }
        phone:         { type: string }
        name:          { type: string }
        nationalId:    { type: string }
        county:        { type: string }
        farmSizeAcres: { type: number }
        cropType:      { type: string }
        createdAt:     { type: string }
    FarmerInput:
      type: object
      required: [phone, name, nationalId, county, farmSizeAcres, cropType]
      properties:
        phone:         { type: string }
        name:          { type: string }
        nationalId:    { type: string }
        county:        { type: string }
        farmSizeAcres: { type: number }
        cropType:      { type: string }

    # ─── Loan Application ────────────────────────────────────────────────────
    LoanApplication:
      type: object
      required: [id, farmerId, amountRequested, purpose, status, createdAt]
      properties:
        id:              { type: integer }
        farmerId:        { type: integer }
        farmerName:      { type: ["string", "null"] }
        farmerPhone:     { type: ["string", "null"] }
        amountRequested: { type: number }
        purpose:         { type: string }
        status:          { type: string }
        aiScore:         { type: ["number", "null"] }
        aiRecommendation: { type: ["string", "null"] }
        reviewNote:      { type: ["string", "null"] }
        createdAt:       { type: string }
        updatedAt:       { type: string }
    LoanApplicationInput:
      type: object
      required: [farmerId, amountRequested, purpose]
      properties:
        farmerId:        { type: integer }
        amountRequested: { type: number }
        purpose:         { type: string }
    LoanDecision:
      type: object
      properties:
        reviewNote: { type: string }
        approvedAmount: { type: number }
    CreditScoreResult:
      type: object
      required: [score, recommendation, factors]
      properties:
        score:          { type: number }
        recommendation: { type: string }
        factors:
          type: array
          items:
            type: string

    # ─── Loan ────────────────────────────────────────────────────────────────
...