# lib/api-zod/src/generated/types/farmerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgriPride Mkopo - Agricultural Loan Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FarmerInput {
  phone: string;
  name: string;
  nationalId: string;
  county: string;
  farmSizeAcres: number;
  cropType: string;
}