# lib/api-zod/src/generated/types/loanApplication.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgriPride Mkopo - Agricultural Loan Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LoanApplication {
  id: number;
  farmerId: number;
  /** @nullable */
  farmerName?: string | null;
  /** @nullable */
  farmerPhone?: string | null;
  amountRequested: number;
  purpose: string;
  status: string;
  /** @nullable */
  aiScore?: number | null;
  /** @nullable */
  aiRecommendation?: string | null;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: string;
  updatedAt?: string;
}