# lib/api-zod/src/generated/types/loanDecision.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgriPride Mkopo - Agricultural Loan Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LoanDecision {
  reviewNote?: string;
  approvedAmount?: number;
}